This is my website about dogs
There are 5 files, only 4 of which are used. Assignment.html is the home page, are you are able to navigate to different pages via the nav bar. The two pages you can navigate to are a page including an image of a dog, and a page including a youtube video of a dog.